+++ kernel/classes/datatypes/ezmedia/ezmediatype.php
@@ -109,20 +109,23 @@ class eZMediaType extends eZDataType
         {
             $count = 0;
             $currentBinaryFile =& eZMedia::fetch( $contentObjectAttributeID, $version );
-            $mimeType =  $currentBinaryFile->attribute( "mime_type" );
-            $currentFileName = $currentBinaryFile->attribute( "filename" );
-            list( $prefix, $suffix ) = split ('[/]', $mimeType );
-            $orig_dir = "var/storage/original/" . $prefix;
-            foreach ( $mediaFiles as $mediaFile )
+            if ( $currentBinaryFile != null )
             {
-                $fileName = $mediaFile->attribute( "filename" );
-                if( $currentFileName == $fileName )
-                     $count += 1;
-            }
-            if( $count == 1 )
-            {
-                if( file_exists( $orig_dir . "/" . $currentFileName ) )
-                    unlink( $orig_dir . "/" .  $currentFileName );
+                $mimeType =  $currentBinaryFile->attribute( "mime_type" );
+                $currentFileName = $currentBinaryFile->attribute( "filename" );
+                list( $prefix, $suffix ) = split ('[/]', $mimeType );
+                $orig_dir = "var/storage/original/" . $prefix;
+                foreach ( $mediaFiles as $mediaFile )
+                {
+                    $fileName = $mediaFile->attribute( "filename" );
+                    if( $currentFileName == $fileName )
+                        $count += 1;
+                }
+                if( $count == 1 )
+                {
+                    if( file_exists( $orig_dir . "/" . $currentFileName ) )
+                        unlink( $orig_dir . "/" .  $currentFileName );
+                }
             }
         }
         eZMedia::remove( $contentObjectAttributeID, $version );
